requirements_oauth.txt
-r requirements.txt

django-allauth
django-allauth[socialaccount]
